+++ CHANGELOG.md
@@ -2,6 +2,7 @@ medialibrary CHANGELOG
 ======================
 
 ## CURRENT
+ - [medialibrary-180](https://github.com/cjcodeproj/medialibrary/issues/180) Simple Album classification
  - [medialibrary-179](https://github.com/cjcodeproj/medialibrary/issues/179) Finally need a group noun
  - [medialibrary-171](https://github.com/cjcodeproj/medialibrary/issues/171) Content Index objects should get their own hierarchy
  - [medialibrary-170](https://github.com/cjcodeproj/medialibrary/issues/170) Song support